searchText moved to reducer
